当元素在视窗不可见时，自动滚动到可见区域
