+++ sweetest/sweetest/config.py
@@ -1,3 +1,4 @@
+
 web_keywords = {
     '打开': 'OPEN',
     'OPEN': 'OPEN',
