+++ sweetest/sweetest/testsuite.py
@@ -1,14 +1,17 @@
-from sweetest.globals import g
+import time
+from sweetest.globals import g, timestamp
 from sweetest.windows import w
 from sweetest.testcase import TestCase
 from sweetest.log import logger
 
 
 class TestSuite:
-    def __init__(self, testsuite, report, conditions={}):
+    def __init__(self, testsuite, sheet_name, report, conditions={}):
         self.testsuite = testsuite
+        self.sheet_name = sheet_name
         self.report = report
         self.conditions = conditions
+        self.result = {}
 
         # base 在整个测试套件中首先执行一次
         self.base_testcase = {}
@@ -23,6 +26,19 @@ class TestSuite:
             elif testcase['condition'].lower() == 'snippet':
                 g.snippet[testcase['id']] = testcase
                 testcase['flag'] = 'N'
+
+
+    def testsuite_start(self):
+        self.result['no'] = g.no
+        g.no += 1      
+        self.result['testsuite'] = self.sheet_name  
+        self.result['start_timestamp'] = timestamp()
+
+
+    def testsuite_end(self):
+        self.result['end_timestamp'] = timestamp() 
+        g.testsuite_data[self.sheet_name] =  self.result
+
 
     def setup(self, testcase, case):
         logger.info('Start running the SETUP testcase...')
@@ -63,6 +79,8 @@ class TestSuite:
 
     def run(self):
 
+        self.testsuite_start()
+
         # 当前测试用例
         current = {'result': 'Pass'}
         # 上一个测试用例
@@ -70,7 +88,6 @@ class TestSuite:
 
         # 1.执行用例
         for testcase in self.testsuite:
-
             # 根据筛选条件，把不需要执行的测试用例跳过
             flag = False
             for k,v in self.conditions.items():
@@ -81,14 +98,15 @@ class TestSuite:
                     flag = True
             if flag:
                 continue
-
+            # 统计开始时间
+            testcase['start_timestamp'] = timestamp()
             # xml 测试报告-测试用例初始化
             if testcase['flag'] != 'N':
                 case = self.report.create_case(
                     testcase['title'], testcase['id'])
                 case.start()
                 case.priority = testcase['priority']
-                # 用例上下文，
+                # 用例上下文
                 previous = current
                 current = testcase
             else:
@@ -96,6 +114,8 @@ class TestSuite:
                 # case.skip('Skip', 'Autotest Flag is N')
                 logger.info('Run the testcase: %s|%s Skipped, because the flag=N or the condition=snippet' % (
                     testcase['id'], testcase['title']))
+                # 统计结束时间
+                testcase['end_timestamp'] = timestamp()
                 continue
 
             if testcase['condition'].lower() not in ('base', 'setup'):
@@ -106,6 +126,8 @@ class TestSuite:
                             'Blocked', 'Main or pre Sub testcase is not pass')
                         logger.warn('Run the testcase: %s|%s Blocked, Main or pre Sub TestCase is not Pass' % (
                             testcase['id'], testcase['title']))
+                        # 统计结束时间
+                        testcase['end_timestamp'] = timestamp()    
                         continue
                 # 如果前置条件为 skip，则此用例不执行前置条件
                 elif testcase['condition'].lower() == 'skip':
@@ -114,11 +136,17 @@ class TestSuite:
                     result = self.setup(testcase, case)
                     #if result == 'N':
                     if not result:
+                        # 统计结束时间
+                        testcase['end_timestamp'] = timestamp()
                         continue
 
             try:
                 tc = TestCase(testcase)
                 tc.run()
+
+                # 统计结束时间
+                testcase['end_timestamp'] = timestamp()
+
                 if testcase['result'] == 'Pass':
                     case.succeed()
                 elif testcase['result'] == 'Fail':
@@ -155,3 +183,5 @@ class TestSuite:
                 logger.info('--- Quit th Driver: %s' %g.browserName) 
         except:
             logger.exception('Clear the env is fail')
+
+        self.testsuite_end()
